requirements.txt
# Core dependencies
requests>=2.31.0  # For Azure OpenAI API calls and token authentication
python-dotenv>=1.0.0  # For environment variable management

# A2A Protocol dependencies
a2a-sdk>=0.2.0  # Agent-to-Agent protocol SDK
uvicorn>=0.30.0  # ASGI server for running the A2A server
starlette>=0.38.0  # ASGI framework (used by A2A SDK)
sse-starlette>=2.0.0  # Server-Sent Events for Starlette (required by A2A SDK)

# OAuth2 Authentication
PyJWT>=2.8.0  # JWT token generation and validation
python-multipart>=0.0.9  # Form parsing for token endpoint

# Optional dependencies for future enhancements:
# pydantic>=2.0.0  # For data validation (included via a2a-sdk)
# pytest>=7.4.0  # For testing
# black>=23.0.0  # For code formatting
# mypy>=1.0.0  # For type checking
